Update skeleton.Rmd
Added `papaja::r_refs('r-references.bib')` to the first code chunk. Idea here is that when I try to compile this from scratch, the first time, it warns me that r-references.bib is missing, and if you're using a container to execute this as a one-off (e.g. Code Ocean), it'll be a problem each time. So the point is to create the file before it is called in line 84.